Text-Threader_Backend/requirements.txt
Django~=2.1.4
djangorestframework~=3.9.0
sklearn~=0.0
scikit-learn~=0.20.2